fix(ios): rename BugReporting TurboModule spec to avoid duplicate symbols
React Native core ships Libraries/BugReporting/NativeBugReporting.js, so
FBReactNativeSpec codegen emits NativeBugReportingSpecJSI. Our spec file
had the same basename and emitted the identical class into RNLuciqSpec,
producing 3 duplicate symbols at link time (test_ios_spm).

Codegen derives generated class names from the spec filename, so renaming
it to NativeLuciqBugReporting.ts makes the symbols unique. The Android
old-arch shim is renamed to match the new-arch generated superclass.

The runtime module name is unchanged (LCQBugReporting in both
TurboModuleRegistry.get and RCT_EXPORT_MODULE), so this is not a breaking
change. iOS needs no edit: LuciqBugReportingBridge is a plain
RCTEventEmitter <RCTBridgeModule> and never conformed to the generated
protocol.

Verified by running codegen locally: RNLuciqSpec and FBReactNativeSpec
symbol sets now have zero overlap (previously 3).
